Move SyncData into its own file
Minor refactoring in preparation for upcoming changes.
